When facing an empty vocab list, the generator throws an error

+++ src/generator/ArtifactGenerator.js
@@ -9,6 +9,9 @@ const VocabGenerator = require('./VocabGenerator');
 const ARTIFACT_DIRECTORY_ROOT = '/Generated';
 const ARTIFACT_DIRECTORY_SOURCE_CODE = `${ARTIFACT_DIRECTORY_ROOT}/SourceCodeArtifacts`;
 
+// This error is thrown when the list of vocabularies in the provided config is empty
+class NoVocabularyProvidedError extends Error {}
+
 class ArtifactGenerator {
   constructor(argv, inquirerProcess) {
     this.artifactData = argv;
@@ -117,6 +120,15 @@ class ArtifactGenerator {
     // Provide access to our entire YAML data.
     this.artifactData.generationDetails = generationDetails;
 
+    // If the vocab list is non-existent or empty (e.g. after initialization), the generator
+    // cannot run.
+    if (!generationDetails.vocabList) {
+      throw new NoVocabularyProvidedError(
+        'No vocabularies found: nothing to generate. ' +
+          'Please edit the YAML configuration file to provide vocabularies to generate from.'
+      );
+    }
+
     // For each programming language artifact we generate, first clear out the destination directories.
     const directoryDeletionPromises = generationDetails.artifactToGenerate.map(artifactDetails => {
       return ArtifactGenerator.deleteDirectory(
@@ -124,12 +136,6 @@ class ArtifactGenerator {
       );
     });
     await Promise.all(directoryDeletionPromises);
-
-    // If the vocab list is non-existant of empty (e.g. after initialization), the generator
-    // cannot run.
-    if (generationDetails.vocabList === undefined || generationDetails.vocabList === null) {
-      return this.artifactData;
-    }
 
     // TODO: This code evolved from where we originally only had a list of
     //  vocabs to generate from. But now we can create artifacts for multiple
@@ -208,5 +214,6 @@ class ArtifactGenerator {
 }
 
 module.exports = ArtifactGenerator;
+module.exports.NoVocabularyProvidedError = NoVocabularyProvidedError;
 module.exports.ARTIFACT_DIRECTORY_ROOT = ARTIFACT_DIRECTORY_ROOT;
 module.exports.ARTIFACT_DIRECTORY_SOURCE_CODE = ARTIFACT_DIRECTORY_SOURCE_CODE;
